dev: Added Client class and binding managers
Added client java class and its associated data access objects and managers.
This client version is very simple: it has an idea which will be 'car1', 'car2',
etc. And a Double attribute which corresponds to the factor by which the ticket
fee is reduced. Now I have to create a custom version of the ticket.